+++ package.json
@@ -58,7 +58,7 @@
   "dependencies": {
     "@babel/runtime": "7.x",
     "@vant/icons": "1.3.0",
-    "@vant/use": "^0.0.3",
+    "@vant/use": "^0.0.4",
     "vue-lazyload": "1.2.3"
   },
   "peerDependencies": {
