+++ packages/vant-use/changelog.md
@@ -1,5 +1,11 @@
 # 更新日志
 
+### v0.0.4
+
+`2020-10-05`
+
+- 新增 `useRect` 方法
+
 ### v0.0.3
 
 `2020-09-27`
